fixed all errors init works
