+++ wsgi.py
@@ -5,6 +5,7 @@ from flask.cli import with_appcontext, AppGroup
 from App.database import db, get_migrate
 from App.main import create_app
 from App.controllers import ( create_user, get_all_users_json, get_all_users )
+from App.controllers.user import initialize_db
 
 # This commands file allow you to create convenient CLI commands for testing controllers
